requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-core==0.3.0
langchain-google-genai>=2.0.0
google-generativeai>=0.8.3
duckduckgo-search
textblob
python-dotenv
ddgs